Enhancements in code architecture
