+++ config/webpack.dev.js
@@ -16,7 +16,7 @@ const tmxVenus = {
     },
   },
   plugins: [new HtmlWebpackPlugin({
-    template: './src/index.html'
+    template: './public/index.html'
   })],
 
 }
