--- deploy.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-mkdir build
-cp ./src/index.html ./src/index.js ./src/css/style.min.css build/
-cp -r ./src/images/ build/
