+++ package.json
@@ -10,11 +10,9 @@
   },
   "devDependencies": {
     "@babel/core": "^7.17.5",
-    "@babel/plugin-transform-modules-commonjs": "^7.16.8",
     "@babel/preset-env": "^7.16.11",
     "@webpack-cli/init": "^1.1.3",
     "babel-loader": "^8.2.3",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "css-loader": "^6.7.1",
     "eslint": "^8.11.0",
     "eslint-config-airbnb-base": "^15.0.0",
@@ -33,7 +31,7 @@
     "webpack-dev-server": "^4.7.4"
   },
   "scripts": {
-    "server": "json-server --watch db.json",
+    "server": "json-server --watch ./config/db.json",
     "prestart": "webpack --config ./config/webpack.dev.js",
     "start": "webpack serve --open --config ./config/webpack.dev.js & npm run server",
     "prebuild": "eslint ./src/* --fix",
