+++ public/index.html
@@ -10,7 +10,7 @@
             href="https://pro.fontawesome.com/releases/v5.10.0/css/all.css" 
             integrity="sha384-AYmEC3Yw5cVb3ZcuHtOA93w35dYTsvhLPVnYs9eStHfGJvOvKxVfELGroGkvsg+p" 
             crossorigin="anonymous"/>
-        <link rel="icon" href="./images/favicon.png">
+        <link rel="icon" href="../src/images/favicon.png">
 
     <body>
         <div class="container">
