+++ src/modules/elements.js
@@ -1,0 +1,8 @@
+const ELEM = {
+    VALUEITEM: document.querySelector('.value-item'),
+    ADDITEM: document.querySelector('.add-item'),
+    ALLITEMS: document.querySelector('.all-items'),
+};
+
+export { ELEM };
+
