+++ src/modules/index.js
@@ -1,0 +1,8 @@
+import { api, fetchAPI } from './api';
+import { ELEM }  from './elements';
+
+export {
+    api, 
+    fetchAPI,
+    ELEM
+}
