+++ src/services/eventlistener.js
@@ -1,0 +1,3 @@
+const eventListener = (el, event, fn) => el.addEventListener(event, fn);
+
+export { eventListener }
