+++ src/services/index.js
@@ -1,0 +1,3 @@
+import { eventListener } from "./eventlistener";
+
+export { eventListener }
